refactor: split model host dependencies
