+++ Memora/Core/Models/GoogleMeetSettings.swift
@@ -30,25 +30,6 @@ final class GoogleMeetSettings {
         self.isEnabled = isEnabled
     }
 
-    // MARK: - Token State
-
-    var isTokenValid: Bool {
-        let accessToken = KeychainService.load(key: .googleMeetAccessToken)
-        guard !accessToken.isEmpty,
-              let expiresAt = KeychainService.loadDate(key: .googleMeetTokenExpiresAt) else {
-            return false
-        }
-        return expiresAt > Date()
-    }
-
-    var shouldRefreshToken: Bool {
-        let refreshToken = KeychainService.load(key: .googleMeetRefreshToken)
-        guard !refreshToken.isEmpty else { return false }
-        guard let expiresAt = KeychainService.loadDate(key: .googleMeetTokenExpiresAt) else { return true }
-        // 5 分前にリフレッシュ
-        return expiresAt.addingTimeInterval(-300) <= Date()
-    }
-
     // MARK: - OAuth Configuration
 
     /// 認可エンドポイント
