+++ Memora/Core/Models/TranscriptionCheckpoint.swift
@@ -22,51 +22,3 @@ final class TranscriptionCheckpoint {
         self.chunkResultsBlob = Data()
     }
 }
-
-/// チェックポイントに保存するチャンク結果。
-/// TranscriptionResult 自体を Codable 化せず、保存用の独立 DTO とする。
-struct CheckpointChunkResult: Codable, Sendable {
-    struct Segment: Codable, Sendable {
-        let id: String
-        let speakerLabel: String
-        let startSec: Double
-        let endSec: Double
-        let text: String
-        let isEstimatedTiming: Bool
-    }
-    let fullText: String
-    let language: String
-    let segments: [Segment]
-
-    init(from result: TranscriptionResult) {
-        self.fullText = result.fullText
-        self.language = result.language
-        self.segments = result.segments.map {
-            Segment(
-                id: $0.id,
-                speakerLabel: $0.speakerLabel,
-                startSec: $0.startSec,
-                endSec: $0.endSec,
-                text: $0.text,
-                isEstimatedTiming: $0.isEstimatedTiming
-            )
-        }
-    }
-
-    func toTranscriptionResult() -> TranscriptionResult {
-        TranscriptionResult(
-            fullText: fullText,
-            language: language,
-            segments: segments.map {
-                TranscriptionSegment(
-                    id: $0.id,
-                    speakerLabel: $0.speakerLabel,
-                    startSec: $0.startSec,
-                    endSec: $0.endSec,
-                    text: $0.text,
-                    isEstimatedTiming: $0.isEstimatedTiming
-                )
-            }
-        )
-    }
-}
